Address Philipp and Robert's comments
